requirements.txt
pip packages:
allure-pytest-bdd        2.8.35
allure-python-commons    2.8.35
appdirs                  1.4.4
atomicwrites             1.4.0
attrs                    20.3.0
certifi                  2020.12.5
chardet                  4.0.0
colorama                 0.4.4
distlib                  0.3.1
enum34                   1.1.10
filelock                 3.0.12
glob2                    0.7
idna                     2.10
iniconfig                1.1.1
lxml                     4.6.2
Mako                     1.1.4
MarkupSafe               1.1.1
namedlist                1.8
packaging                20.9
parse                    1.19.0
parse-type               0.5.2
pip                      21.1.1
pluggy                   0.13.1
py                       1.10.0
pyparsing                2.4.7
pytest                   6.2.2
pytest-base-url          1.4.2
pytest-bdd               4.0.2
pytest-failed-screenshot 1.0.0
pytest-html              3.1.1
pytest-metadata          1.11.0
pytest-selenium          2.0.1
pytest-variables         1.9.0
requests                 2.25.1
selenium                 3.141.0
setuptools               49.2.1
six                      1.15.0
tenacity                 6.3.1
toml                     0.10.2
urllib3                  1.26.3
virtualenv               20.4.2

pytest-bdd~=4.0.2
pytest~=6.2.2
selenium~=3.141.0